FACES-3132 Support ExternalContext.setResponseStatus(int statusCode)
