+++ bridge-impl/src/main/java/com/liferay/faces/bridge/context/internal/ExternalContextCompat_Portlet3_Impl.java
@@ -18,6 +18,7 @@ package com.liferay.faces.bridge.context.internal;
 import javax.portlet.PortletContext;
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
+import javax.portlet.ResourceResponse;
 import javax.portlet.faces.Bridge;
 
 
@@ -33,7 +34,35 @@ public abstract class ExternalContextCompat_Portlet3_Impl extends ExternalContex
 		super(portletContext, portletRequest, portletResponse);
 	}
 
-	// Java 1.6+ @Override
+	/**
+	 * Sets the status of the portlet response to the specified status code. Note that this is only possible for a
+	 * portlet request of type PortletResponse because that is the only type of portlet response that is delivered
+	 * directly back to the client (without additional markup added by the portlet container).
+	 *
+	 * @see    {@link ExternalContext#setResponseStatus(int)}
+	 * @since  JSF 2.0
+	 */
+	@Override
+	public void setResponseStatus(int statusCode) {
+
+		if (portletResponse instanceof ResourceResponse) {
+
+			ResourceResponse resourceResponse = (ResourceResponse) portletResponse;
+			resourceResponse.setProperty(ResourceResponse.HTTP_STATUS_CODE, Integer.toString(statusCode));
+		}
+		else {
+
+			if (manageIncongruities) {
+				incongruityContext.setResponseStatus(statusCode);
+			}
+			else {
+				// Mojarra will call this method if a runtime exception occurs during execution of the JSF lifecycle, so
+				// must not throw an IllegalStateException.
+			}
+		}
+	}
+
+	@Override
 	protected boolean isHeaderPhase(Bridge.PortletPhase portletPhase) {
 		return false;
 	}
